# project/backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RepoPulse API scaffold",
    "version": "0.1.0",
    "description": "Current runtime contract for task-17661. Domain endpoints are placeholders. Auth and successful domain responses will be specified when implemented. GET endpoints also support HEAD. Unknown paths return JSON 404; unsupported methods return JSON 405 with Allow."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Check that the API process is alive",
        "responses": {
          "200": {
            "description": "The HTTP process is running. No dependency checks are performed.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "post": {
        "operationId": "startAnalysis",
        "summary": "Reserved: start analysis",
        "description": "Scaffold only. Always returns 501; does not parse or validate the request body. The planned request uses repository_url.",
        "responses": {
          "501": {
            "description": "Analysis orchestration has not been implemented.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "operationId": "getAnalysisStatus",
        "summary": "Reserved: get analysis status",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Analysis status retrieval has not been implemented.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "operationId": "getReport",
        "summary": "Reserved: get report",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Report retrieval has not been implemented.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "headers": {
      "RequestID": {
        "description": "A new server-generated request ID; also written to HTTP request logs.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "Health": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "NOT_IMPLEMENTED",
                  "NOT_FOUND",
                  "METHOD_NOT_ALLOWED"
                ]
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}